# trie: db insertions + leaf callback parallelism