--- a/trie/hasher.go
+++ b/trie/hasher.go
@@ -184,7 +184,7 @@ func (h *hasher) store(n node, db *Database, force bool) (node, error) {
 		hash := common.BytesToHash(hash)
 
 		db.lock.Lock()
-		db.insert(hash, h.tmp, n)
+		db.insert(hash, len(h.tmp), n)
 		db.lock.Unlock()
 
 		// Track external references from account->storage trie