When we do `Commit` on the trie when parents-of-leafs are encountered, we do two things: 

1. call `db.insert`, which can be a fairly heavy operation, and 
2. call the `onleaf` function for the leafs. This can also be somewhat heavy. 

This PR splits this up, so that we can have one dedicated goroutine which handles the insertion and callbacks, and the main trie-walker can concentrate on iterating the trie. 

Since the order of events (inserts and callback invocation) is unchanged by this PR, I don't think this has any buggy side-effects.

Here are the benchmark results: 
```
name                           old time/op    new time/op    delta
CommitAfterHash/no-onleaf-6      5.10µs ±14%    4.28µs ± 8%  -16.06%  (p=0.000 n=10+9)
CommitAfterHash/with-onleaf-6    6.77µs ±19%    5.44µs ±18%  -19.71%  (p=0.000 n=10+10)

name                           old alloc/op   new alloc/op   delta
CommitAfterHash/no-onleaf-6      1.79kB ± 3%    1.94kB ± 5%   +8.65%  (p=0.000 n=10+10)
CommitAfterHash/with-onleaf-6    2.18kB ± 2%    2.22kB ± 2%   +1.81%  (p=0.004 n=10+10)

name                           old allocs/op  new allocs/op  delta
CommitAfterHash/no-onleaf-6        11.0 ± 0%      13.0 ± 0%  +18.18%  (p=0.000 n=10+10)
CommitAfterHash/with-onleaf-6      18.3 ± 4%      20.0 ± 0%   +9.29%  (p=0.000 n=10+10)

```
As can be expected, if there's an `onleaf` function provided, the gains are higher `20%`, compared to `16%` if there is no `onleaf`. 
The benchmarks are pretty synthetic, though, and I imagine that there is a lot more work done in the `insert` section on a live node. 

This PR should work pretty nicely even pre-byzantium, I'd guess. 

The actual implementation can surely be cleaned up a bit, particularly the hack in `trie.go` could be made a bit more elegant.