@holiman 2019-12-18
I added a benchmark on fixed-size tries, to measure what the impact is on smaller tries. 
```
BenchmarkCommitAfterHashFixedSize/10-6         	   16898	     71678 ns/op	   41125 B/op	     270 allocs/op
BenchmarkCommitAfterHashFixedSize/100-6        	    3934	    278342 ns/op	  189006 B/op	    1292 allocs/op
BenchmarkCommitAfterHashFixedSize/1000-6       	     457	   2592712 ns/op	 1917490 B/op	   13264 allocs/op
BenchmarkCommitAfterHashFixedSize/10000-6      	      27	  41369979 ns/op	20820080 B/op	  135751 allocs/op
```

Before:
```
BenchmarkCommitAfterHashFixedSize/10-6             20391             59862 ns/op           37498 B/op        240 allocs/op
BenchmarkCommitAfterHashFixedSize/100-6             4272            254688 ns/op          178676 B/op       1157 allocs/op
BenchmarkCommitAfterHashFixedSize/1000-6             403           2910285 ns/op         1828260 B/op      11901 allocs/op
BenchmarkCommitAfterHashFixedSize/10000-6             28          42699494 ns/op        19927566 B/op     121839 allocs/op
```

The channel-based `Commit` is slower for trie sizes of `<1000` items, so it would be good to retain the existing non-channel-based `Commit` for smaller tries. Either by having a measure of the number of changes in the trie (adding a counter to `Update`) or simply use a different method when invoking account commit versus storage commit.