@holiman 2019-12-18
I've put this up on a full sync benchmark against master: https://geth-bench.ethdevops.io/d/Jpk-Be5Wk/dual-geth?orgId=1&var-exp=mon08&var-master=mon09&var-percentile=50&from=1576664098483&to=now