@holiman 2019-12-19
Full sync in progress, it's just passed byzantium at 4.37M. 
![Screenshot_2019-12-19 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/71155751-b2b8e580-223e-11ea-943e-b35d4e87e399.png)

The exp machine is, I believe, a bit slower than master, and the exp is a bit behind in terms of blocks (seems to be about ten minutes). However, looking at the account commit chart, it seems that this PR is a couple of of milliseconds (or, maybe 20-25%) faster than master per block.