@holiman 2019-12-19
This PR hit byzantium about 5 minutes after `master`. After byzantium (`4.37M`), it started inching closer to master, and around block `4.822M` it overtook it. 
![Screenshot_2019-12-19 Dual Geth - Grafana(1)](https://user-images.githubusercontent.com/142290/71195529-67c6be80-228e-11ea-8f42-280b395a4d00.png)

This chart shows the account commit: 
![Screenshot_2019-12-19 Dual Geth - Grafana(2)](https://user-images.githubusercontent.com/142290/71195655-ad838700-228e-11ea-8a98-7fa2e3e7cae6.png)

During the high-intensity periods after byz, master peaks at `26ms` , while this one peaks at `16ms`.