@holiman 2019-12-21
Here's the last two days: 
![Screenshot_2019-12-21 Dual Geth - Grafana](https://user-images.githubusercontent.com/142290/71305031-774d2100-23ce-11ea-8dfd-e7956bcad0ea.png)
From block ~`4.56M` to `6.02M`. Experimental leads by about 10 minutes at the end.