@holiman 2019-12-23
After about 5 days of sync, the experimental is ahead with about `35` minutes. They're at block `7.6M`, and I'm about to replace the images with another experiment.