@holiman 2020-01-04
Closing in favour of https://github.com/ethereum/go-ethereum/pull/20481